refactor: move TerminationStatus converter
